PA_CVAE/condaenv.dx7vx0d0.requirements.txt
absl-py==2.0.0
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.2
charset-normalizer==3.3.2
colorama==0.4.6
cycler==0.11.0
datetime==5.2
decorator==5.1.1
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.19.1
joblib==1.3.2
keras==2.10.0
keras-applications==1.0.8
keras-preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
markdown==3.4.4
markupsafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
pillow==9.5.0
pip==23.3.1
prompt-toolkit==3.0.43
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-graphviz==0.20.1
python-version==0.0.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.66.1
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
zope-interface==6.1